+++ lib/departure/rails_adapter.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
+require 'forwardable'
+
 module Departure
   class RailsAdapter
-    extend Forwardable
+    extend ::Forwardable
 
     class << self
       def version_matches?(version_string, compatibility_string = current_version::STRING)
-        raise "Invalid Gem Version: '#{version_string}'" unless Gem::Version.correct?(version_string)
-
         requirement = Gem::Requirement.new(compatibility_string)
         requirement.satisfied_by?(Gem::Version.new(version_string))
       end
@@ -21,7 +21,9 @@ module Departure
       end
 
       def for(ar_version)
-        if ar_version::MAJOR == 8
+        if ar_version::MAJOR == 8 && ar_version::MINOR.positive?
+          V8_1_Adapter
+        elsif ar_version::MAJOR == 8
           V8_0_Adapter
         elsif ar_version::MAJOR >= 7 && ar_version::MINOR >= 2
           V7_2_Adapter
@@ -77,6 +79,16 @@ module Departure
             connection_options,
             config
           )
+        end
+
+        # https://github.com/rails/rails/commit/9ad36e067222478090b36a985090475bb03e398c#diff-de807ece2205a84c0e3de66b0e5ab831325d567893b8b88ce0d6e9d498f923d1
+        # Rails Column arity changed to require cast_type in position 2 which required us introducing this indirection
+        def new_sql_column(name:,
+                           default_value:,
+                           mysql_metadata:,
+                           null_value:,
+                           **_kwargs)
+          sql_column.new(name, default_value, mysql_metadata, null_value)
         end
 
         def sql_column
@@ -139,6 +151,48 @@ module Departure
 
         def sql_column
           ::ActiveRecord::ConnectionAdapters::Rails80DepartureAdapter::Column
+        end
+      end
+    end
+
+    class V8_1_Adapter < BaseAdapter # rubocop:disable Naming/ClassAndModuleCamelCase
+      class << self
+        def register_integrations
+          require 'active_record/connection_adapters/rails_8_1_departure_adapter'
+          require 'departure/rails_patches/active_record_migrator_with_advisory_lock_patch'
+
+          ActiveSupport.on_load(:active_record) do
+            ActiveRecord::Migration.class_eval do
+              include Departure::Migration
+            end
+
+            ActiveRecord::Migrator.prepend Departure::RailsPatches::ActiveRecordMigratorWithAdvisoryLockPatch
+          end
+
+          ActiveRecord::ConnectionAdapters.register 'percona',
+                                                    'ActiveRecord::ConnectionAdapters::Rails81DepartureAdapter',
+                                                    'active_record/connection_adapters/rails_8_1_departure_adapter'
+        end
+
+        def create_connection_adapter(**config)
+          ActiveRecord::ConnectionAdapters::Rails81DepartureAdapter.new(config)
+        end
+
+        # rubocop:disable Metrics/ParameterLists
+        # https://github.com/rails/rails/commit/9ad36e067222478090b36a985090475bb03e398c#diff-de807ece2205a84c0e3de66b0e5ab831325d567893b8b88ce0d6e9d498f923d1
+        # Rails Column arity changed to require cast_type in position 2 which required us introducing this indirection
+        def new_sql_column(name:,
+                           cast_type:,
+                           default_value:,
+                           mysql_metadata:,
+                           null_value:,
+                           **_kwargs)
+          # rubocop:enable Metrics/ParameterLists
+          sql_column.new(name, cast_type, default_value, mysql_metadata, null_value)
+        end
+
+        def sql_column
+          ::ActiveRecord::ConnectionAdapters::MySQL::Column
         end
       end
     end