+++ lib/departure/runner.rb
@@ -1,10 +1,15 @@
 require 'open3'
+require 'forwardable'
+
+# Runner is deprecated and used in rails 8.0 and below
+# We will slowly reduce the responsibility of departure to be delegating
+# all calls to underlying adapters/connections
 
 module Departure
   # It executes pt-online-schema-change commands in a new process and gets its
   # output and status
   class Runner
-    extend Forwardable
+    extend ::Forwardable
 
     def_delegators :raw_connection, :execute, :escape, :close, :affected_rows